- Traverse the trees prior to code generation t...
- Traverse the trees prior to code generation to determine the class
symbols for which .NET types should be generated. Also detect the entry
point of the program. If more entry points are encountered issue a
warning for each of them and set the first one as the real entry point -
Determine the name of the output assembly from (1) command-line option
(2) the name of the class that defines the entry point or (3) the name
of the first sourcefile