- added allowUnsafeArrays thread-local variable...
- added allowUnsafeArrays thread-local variable in Type, to allow or
disallow the creation of "unsafe" arrays by the current thread. Unsafe
arrays are defined as those whose type, after erasure, is the same as
the erasure of some other array type. (For example, an array of "C
with D" is indistinguishable from an array of "C" after erasure, hence
unsafe).