fixed bug with type alias expansion that turned...
fixed bug with type alias expansion that turned up when an old-style type constructor alias (e.g. type Class = java.lang.Class in Predef) was compiled without the new checks (in the example, without -Xgenerics), and now it's used with  the checks (with -Xgenerics). Then, normalize didn't do anything (as the number of type params didn't coincide with the number of type arguments). I.e., it returned `this', which resulted in an infinite recursion.