Generalizes the allowed inputs to the interpret...
Generalizes the allowed inputs to the interpreter to be any combination
of phrases that can be placed inside a template. Instead of having one
"request" object try to understand the entire combination of phrases,
the code now uses multiple "member handlers" for each request, one
member handler for each phrase.