The great library reorg!
The standard library will now be laid out approximately like this

 library/scala
        /jvm/scala
        /jvm/jvm1.4/scala
        /jvm/jvm1.5/scala
        /jvm/android/scala
        /jvm/cldc/scala
        /dotnet/scala

To build the standard library for a given target you start by copying
the root library/scala tree to a staging area, then move down in the
hierarchy copying the nested "scala" trees on top in the staging area.
So if you wanted to build for cldc, for example, you would do something
like the following:

  rsync -avz library/scala                staging/
  rsync -avz library/scala/jvm/scala      staging/
  rsync -avz library/scala/jvm/cldc/scala staging/

The ant build files will be updated to do this for you automagically,
and there will soon be shell script to do this if you want to compile
some parts of the standard library manually.