fsc watchdog removal.
In addition, fixed a potential initialization problem in the fsc server,
which would open a socket, get a port number, and close the socket, only
to try to reallocate the same port number a bit later, even though the
port might have become unavailable in the meantime. Now the socket is
initialized only once, and the port is not released.