Modifications to pattern matching to allow the ...
Modifications to pattern matching to allow the threading of a Typer all
the way through. This is in aid of fixing bug #1087.

It doesn't actually fix it, but it makes the error message produces a
little more enlightening. That's something. :)

(It also seems this is a neccessary change for correctness with respect
to the typechecker anyway).