Trying to bring BigDecimal all the way up to sp...
Trying to bring BigDecimal all the way up to speed with its java
counterpart. Added a MathContext val and perform all operations with
that MathContext, and an apply method which accepts a math context so
you can e.g.:

  val x = BigDecimal(1.0)
  x / 3             // ArithmeticException
  x(DECIMAL32) / 3  // 0.3333333

Also added most of the java BigDecimal methods which for whatever reason
weren't exposed by the wrapper.

Precision to infinity... and beyond!