The sort of pattern matcher commit which makes ...
The sort of pattern matcher commit which makes water taste better and
birds sing more sweetly. Hard to be more specific than that, but let's
just say that the number of IQ points one has to have on hand to grok
the pattern matcher is down at least 25 from this time last year.