Finished the additions to RuleApplication.
bits into the SequencePattern class, which took me an eternity because
the current logic only worked because of bugs in the pattern matcher
which allowed a non-matching case to match anyway. Trust me, you do not
want to debug that sort of thing. I wonder how much of the compiler
would stop working if I magically fixed all the matcher bugs at once,
and long dormant logic suddenly lurched to life...