In the same way that raw Trees are giving way t...
In the same way that raw Trees are giving way to Patterns, raw Symbols
(and accompanying ValDef trees) are being replaced by PatternVars. Once
these are both used consistently a number of bugs will vanish with a
puff of smoke, as they are caused by inconsistencies which need not ever
be allowed to occur.