Some reworking of command line option parsing b...
Some reworking of command line option parsing because it was not
possible to say

  scala -i include1.scala include2.scala -e 'code to execute'

and it should be. And now it is. This required changing
MultiStringSetting to only accept arguments until it runs into one
starting with '-' rather than blindly swallowing all the rest of the
command line as it used to, which was terribly unfriendly parsing
behavior regardless.