Some structural improvements to Either and Opti...
Some structural improvements to Either and Option which leverage recent
awesomeness in constraining type parameters. In Either I was able
to define joinLeft and joinRight on the instance rather than on the
object, and while I didn't manage that directly with merge, it can at
least be accomplished via implicit as endorsed by martin 25/Jun/09 on
scala-internals.