Modified partialMap to take PartialFunction[A, ...
Modified partialMap to take PartialFunction[A, B] instead of Any => B.
There was definitely some reason I didn't do this in the first place,
but either that reason no longer applies or it's not easily induced.
However please consider this change slightly tentative as I feel like
that other shoe may still be out there and we might have to reconsider.