Fix for #2365.
a soft reference around cached classes so they do not interfere with
garbage collection. There is a test case, but it is in pending because I
spent longer trying to get it to fail under partest than I did writing
the actual patch. If you would like to see the behavior which was
corrected, go to

  test/pending/run/bug2365

and run that script with scalac built before and after this commit.
Review by dubochet.