One of those "$.05 for the bolt, $50,000 for kn...
One of those "$.05 for the bolt, $50,000 for knowing where to put it"
commits. Closes #425, #816, #2310, #2691. All credit for this patch goes
to me for having the genius to know when new eyes were needed (although
if you're feeling generous some could also go to walter korman for the
actual debugging and code writing part.)