I'm sure I'm not the only one driven into parox...
I'm sure I'm not the only one driven into paroxysms of annoyance at the
fact that repl transcripts are not at all usable in the repl. No longer:
now you can paste a transcript into the repl and it will automatically
detect it as such, clean it up and replay it. It is triggered by the
"scala> " on the first line of the transcript. Review by community.