The remainder of isolating MSIL from the rest o...
The remainder of isolating MSIL from the rest of the classpath code. To
accomplish this I made ClassRep an inner class of ClassPath (which given
the broad definition of ClassPath already in place, it conceptually is
already) and as a bonus this allowed dropping its type parameter. Review
by rytz.