Until now directories on the classpath were not...
Until now directories on the classpath were not considered for repl
completion because of the risk of accidentally traversing large chunks
of the filesystem (e.g. "." in the path, run from /). Some low hanging

fruit was available - at least SCALA_HOME can be considered safe, and
then we get the scala classes.

The main impact of this is that completion now works for the built-in
classes when you run build/quick/bin/scala. Review by community.