Disabled JavaInteraction test.
of time because it fails if you can't connect to the screen of the test
machine. And then if any test fails, the stability test doesn't run. We
badly a separate testing area for tests which are prone to failure for
reasons which are unrelated to the quality ostensibly being tested. No
review.