It turns out some of the weirdness lately is be...
It turns out some of the weirdness lately is because changes to
classpath handling have a way of not taking effect until they're
installed via starr, and presently we have a starr with different logic
than trunk. No choice but to roll up one more starr based on r20984. No
review.