Quite a lot more work on XML equality than I ca...
Quite a lot more work on XML equality than I can properly justify
spending time on, but you know how it is once you get started on
something. This commit will likely break some code out there in the
world but this is impossible to avoid if we are to achieve sane equality
in trunk.

For anyone who was relying upon the 2.7 equality behavior for
scala.xml.* classes, using "xml_==" instead of "==" for comparisons will
restore the old behavior. The standard == on xml elements now attempts
to behave in such a way that symmetry and hashCode contracts will be
preserved. It's probably not 100% there yet, but I can tell you this: it
is closer today than it was yesterday.

Review by community.