Some minor changes in scala.swing.* which I was...
Some minor changes in scala.swing.* which I was glancing through because
of #3196. I noticed the Font object was in package scala instead of
scala.swing, which looks sure to be a mistake (an easy one to make, and
one others have made as well, because we're not entirely used to package
objects.) I didn't want to accidentally ship a scala.Font so I moved it
into swing. Review by imaier.