Altered classpath behavior when no default is g...
Altered classpath behavior when no default is given. Now in that case
the contents of environment variable CLASSPATH will be used as the
scala user classpath, and only if that is not present will "." be used.
Be advised that there are still various "hand assembled" sorts of
classpaths in trunk, and there's not yet any way to ensure they honor
this; things which use the normal Settings object should do the right
thing. No review.