Mostly IO tweaks related to my upcoming partest...
Mostly IO tweaks related to my upcoming partest patch, which to my
chagrin is being held up by windows. Also updates the default ANT_OPTS
to be the same as the ones the nightlies override it with. (If we know
you can't build scala with those settings it seems kind of uncool to
leave them for everyone else.) No review.