Restores the passing of scalac.args and javac.a...
Restores the passing of scalac.args and javac.args from ant to partest
for use on all tests, which rytz had astutely pointed out I made
disappear. Long term I think one ought to be able to specify different
arguments for building the compiler and for partest, but now is not the
moment for such grand ambition. No review.