StringBuilder no longer violates the Seq revers...
StringBuilder no longer violates the Seq reverse contract: it returns
a new StringBuilder. The behavior formerly found in reverse (updates
in place) is now available in reverseContents. Migration warning on
reverse. Closes #3327.

Also did some StringBuilder rewriting as per discussion with odersky.
And took a cleaver to parts of the documentation to get to the good
parts a little faster. Review by community.