Added classes/continuations-plugin to the plugi...
Added classes/continuations-plugin to the plugin build classpath.
Explanation: as things stood the plugin step of the overall build fails
constantly even when no files have been touched. The reason is that if
one checks out an older version of the repository and then returns to
the master, a subset of continuation source files will have more recent
modification dates than their corresponding classes, and ant sees that
as a reason to rebuild them. But without continuations-plugin on the
classpath, it cannot see the classfiles of those which were not changed.

IOW, if a project has A.scala B.scala C.scala and tries to rebuild
only A.scala and C.scala, B.class must be somewhere it can be seen. We
can resolve this differently if desired but this is how the rest of
the compiler does it. (Try removing quick from quick's classpath and
rebuilding after some changes.)

Review by rompf.