Fixed a missed opportunity for code rewriting i...
Fixed a missed opportunity for code rewriting in specialize: a method
that was specialized both on enclosing class types and its own was
routed to the generic one. For instance, a map[@specialized B] would
exhibit this bug.

Restricted warnings to non-synthetic methods, removing warnings for
Tuple default argument methods. No review.