Swapped the order of the arguments to returning...
Swapped the order of the arguments to returning to make it consistent
with the way I've been using it for a long time. No review, how about
a question for nobody in particular instead: can we put this in the
library somewhere so I can stop looking around for it everywhere I go?
It's really handy.

  def returning[T](x: T)(f: T => Unit): T = { f(x) ; x }