Made interpreter robust against a missing class...
Made interpreter robust against a missing classpath and had it issue
informative message in such cases, as per suggestion by Razvan Cojocaru.
No review.

% java -classpath build/pack/lib/'*' scala.tools.nsc.MainGenericRunner
scala> Failed to initialize compiler: class scala.runtime.BooleanRef
not found. ** Note that as of 2.8 scala does not assume use of the java
classpath. ** For the old behavior pass -usejavacp to scala, or if using
a Settings

** object programatically, settings.usejavacp.value = true.