Restored Source factories to a form source comp...
Restored Source factories to a form source compatible with 2.7.7. No
default implicit arguments, now low priority saves the day with a low
priority codec which io.Codec offers as last resort. Dropped the line
separator argument to getLines and made it act in a separator agnostic
way (any of \r\n, \r, or \n is a separator.) Review by community.