Renaming files and inserting dummies to please ...
Renaming files and inserting dummies to please ant. Note to committers:
whenever there is a source file which does not generate a classfile
whose name and path exactly match the name and path of the source file,
then ant will recompile the file every time it compiles anything. In
this batch there were 7 such files, so any 1-char change meant an 8-file
recompile. Someday we'll be rid of ant, but until then... no review.