Now that 2.8.0 is out, I am reverting r22308, w...
Now that 2.8.0 is out, I am reverting r22308, which was the manual patch
that paulp created to address #3420 in a manner that was compatible with
the status of the 2.8.x branch at the time.

After reverting I will re-align 2.8.x with trunk, so the proper fix will
be fetched from there (hopefully).