positioning fix for infix expressions: report e...
positioning fix for infix expressions: report error at the operator
instead of at the receiver of a binop

In general use the operator position as the Select position's point
instead of the left or right expression. Position.union always uses the
receiver's point as point of the result, so swapping fixes that.

See http://www.scala-lang.org/node/6912 for the report.