Revert "CPI -> SCPI to (hopefully) fix scaladoc...
Revert "CPI -> SCPI to (hopefully) fix scaladoc failure in nightly.
probably indicates a bug in scaladoc or the type checker or the spec,
since abstract type members seem to be allowed in objects inconsistently
across these three. no review."

that was a little too hasty.. SCPI was already defined somewhere

This reverts commit 165424cd9175d7e6d33bc541f2aff2575ee31121.