Fix and test case for #3855.
situations where a mutable var will later be lifted. As a point of
interest, this bug reveals itself fairly clearly if you use a build
since r23112 and run the checker thusly:

  scalac -d /tmp -Ycheck-debug -Ycheck:icode -Xprint:icode
test/files/run/bug3855.scala

It dies with the following explanation:

  Output changed for Block 3 [S: 2] [P: 1, 4]
  Exception in thread "main" scala.tools.nsc.backend.icode.CheckerException:
  Incompatible stacks: TypeStack() and TypeStack(2 elems) {
    REFERENCE(class IntRef)
    REFERENCE(class IntRef)
  } in Test.main at entry to block: 2

And indeed that was the source of the reported verifyerror.

Review by i. dragos.