The next batch of tests put up a little more st...
The next batch of tests put up a little more struggle, but only a
little. See test/pending/pos/unappgadteval.scala (the changes for which
were in the previous commit) for an example of a test which might be on
to something. Any idea what it would take to get it working?

  // the key lines
  case i @ Suc()          => { (y: Int) => y + 1 }      // a = Int => Int
  case f @ Lam[b,c](x, e) => { (y: b) => eval(e, env.extend(x, y)) }  // a = b=>c

No review.