Fans of scala cinema who have viewed this types...
Fans of scala cinema who have viewed this types video:
http://www.scala-lang.org/sites/default/files/martin_ordersky_scala_inte
rnals_2009-03-04.avi may have noticed that martin entered a variety of
priceless comments in the course of the lecture and never checked them
in. As the self-appointed royal stenographer I have come forward to
right this wrong.

It being a grainy video and almost inaudible, I may well have butchered
it somewhere, so review by odersky.