An overhaul of the collection-oriented methods ...
An overhaul of the collection-oriented methods in Tuple2/Tuple3 (which
still need to be taken all the way to Tuple5.)

* Zip semantics: zip and zipped now properly terminate when any
collection runs out of elements, even if others are infinite. In
addition, short circuiting methods (exists and forall) will terminate if
the condition is met, even in the presence of infinity.

Example:

  scala> val ys = Stream from 1
  ys: scala.collection.immutable.Stream[Int] = Stream(1, ?)

  scala> (ys, ys).zipped forall ((x, y) => x+y < 100)

  res0: Boolean = false

  scala> (ys, ys).zipped exists ((x, y) => x+y > 100)

  res1: Boolean = true

* There are implicits converting Zipped2/3 to Traversable to expose all
the methods which aren't defined in an arity-specific way in the tuple
classes. I have mixed feelings about putting these in Predef; but if
there is another way to make them visible by default I wasn't able to
find it.

Example putting said implicit to use:

  scala> (ys, ys, ys).zipped find { case (x, y, z) => x+y+z > 1000 }
  res0: Option[(Int, Int, Int)] = Some((334,334,334))

Already reviewed by moors, so no review.