closes #3281, #3866.
I don't understand why we had to clear() the undoLog in the first place,
since the undoXXX methods increase and decrease its size symmetrically,
so the log should always be empty once they have all unwound. Was it a
(premature) optimisation or was there some kind of semantic meaning to
it that I didn't see?

review by odersky