While investigating the cause of #3984, discove...
While investigating the cause of #3984, discovered that set hashcodes
were being calculated like:

  elems map (_.hashCode) sum

This looks totally correct, except that because of set+map semantics it
drops all the duplicate hashcodes and therefore calculates the wrong
sum anytime there are unequal elements with equal hashcodes. If we're
married to the set+map semantics (and I don't know what could be done
about it) I had better get a proper -Xlint going because this is one of
those traps which shoots bees out of its mouth. No review.