Another exciting development in the world of -Y...
Another exciting development in the world of -Y options which I and
three other people will use. Today's is -Yrich-exceptions. Use it like
so:

  SOURCEPATH=/path/to/src scalac -Yrich-exceptions a.scala

In the repl, -Yrich-exceptions will cause lastException to be bound to
an Exceptional instead of old rusty Throwable. That spins up new powers:

  scala> Nil.head
  [Nil.head] (List.scala:389)
  (access lastException for the full trace)
  scala> lastException.show
  /* The repl internal portion of the stack trace is elided. */
  [Nil.head]
   386:   override def isEmpty = true
   387:   override def head: Nothing =
   388:     throw new NoSuchElementException("head of empty list")
  *389:   override def tail: List[Nothing] =
   390:     throw new UnsupportedOperationException("tail of empty list")
   391:   // Removal of equals method here might lead to an infinite recursion similar to IntMap.equals.
   392:   override def equals(that: Any) = that match {

  [line0.<init>] (<console>:6)
  [line0.<clinit>] (<console>:-1)

Also try "lastException.showTable" but this is getting a little long for
more excerpt. No review.