Changing EnclosingMethod generation not to drop...
Changing EnclosingMethod generation not to drop the trailing $ from the
name of the container class. I'm sure this was done for some reason but
proguard objects rather strenuously (and fails) because it can't find
the method, since it says it's in the class but is really in the object.
Review by dragos.