A patch authored by martin to fix one of the bi...
A patch authored by martin to fix one of the big outstanding issues with
-Ycheck. The typer had a vision of the world which fell to pieces if
presented with a tree from after phase constructors. Now it is wiser:
perhaps a bit sadder as well, but every typer must face this day sooner
or later.

Also removed the various passing tests from test/checker-tests, but
still plenty of fascinating failures in there for anyone interested in
helping push to the finish line! No review.