The ability to interrupt repl execution without...
The ability to interrupt repl execution without terminating the JVM has
snowballed a bit, but it's so handy that it's worth it. This commit
further encapsulates individual lines of execution such that they
can be managed after the signal comes in, and if necessary, forcibly
terminated. I'm straying into territory where I could use a really
Thread-knowledgeable reviewer: any volunteers? In the meantime, no
review.