New jline: BSD licensed, based on [ https://git...
New jline: BSD licensed, based on [ https://github.com/jdillon/jline2 ]
with additional code thanks to [ https://github.com/huynhjl/jline2 ].
Replaces lib/jline.jar with build of these sources, and modifies trunk
code to work with the new jar.

Hopeful improvements including baseline functionality on cygwin
and 64bit windows, as well as more accurate line wrapping / cursor
positioning on all platforms and ctrl-R history search.

For the time being the canonical source repository is this:

  https://github.com/paulp/jline2

The enclosed sources are a mirror of that repository, and should be
treated as read-only in the scala svn repository.

No review, codewise, but people are very strongly encouraged to try it
out and report any regressions.