Some profiler changes.
after each phase, so all the allocations are partitionable by phase
in the profiler. I also changed the name of -Yprofile-resident to
-Yprofile-memory and had it snapshot after the run instead of between
them, so it is usable for regular scalac usage as well.

Added tools/profile_scalac which can be used as a drop-in replacement
which handles some tedious setup. Review by dragos.