Brought back and deprecated onlyPresentation to...
Brought back and deprecated onlyPresentation to unbreak sbt. (Another
fine reason for us to push all the way to using sbt is that we wouldn't
unwittingly break it, as that change and my recent jline changes both
did.) Maybe there is a better way to do what's in this one: review by
dragos.