Moved the presentation compiler thread in its o...
Moved the presentation compiler thread in its own top-level class. This
allows to collect the presentation compiler instance when the thread is
not GCed, but the compiler should be discarded (for example, after a
shutdown request). Having the thread as an inner class keeps a reference
to the outer class. review by odersky.