In a code review performance for the ages, I fo...
In a code review performance for the ages, I found an instance of the
bug being fixed in the commit under review still going strong elsewhere
in the file. Since the function with the bug wasn't actually called from
anywhere I fixed it by deleting it. No review could possibly improve
upon that.