More repl housecleaning.
interpreter and the interpreter loop has become so fuzzy it can hardly
be said to exist. In this commit I take a step or two toward clearer
boundaries, and also fix some gratuitous name inconsistencies with the
rest of trunk. A Global should be called "global" in the absence of a
compelling reason otherwise.

Also, now having had a couple years to explore every inch of the design
space, I can announce that where possible to use them, distinguished
objects like NoType and NoSymbol are far superior to swaddling things
in Option. Applied this piece of canon to Completion and History. No
review.