There is a lot of housecleaning to be done.
up the stray interpreter files and put them in the interpreter package.
Would really love to change the name of that package. Went looking for
some consistent divisions of responsibility and consistent naming.
Made some progress. There are deprecated versions of most everything I
changed so hopefully the carnage will be limited.

This isn't completely baked but I just realized I broke the build
earlier and this should fix it. I'll keep the oven on. No review.