After seeing the 481 line diff I had to apply t...
After seeing the 481 line diff I had to apply to bring the checkfile up
to date, I concluded this test cannot be in the general pool: this is
the third build break in the last few days, because it changes output
like some people change socks. (Hygenic people.) Sorry to move it to
disabled iulian, but I don't know what else to do with it on short
notice. Review by dragos.