Fixed all the forms of process input/output red...
Fixed all the forms of process input/output redirection so the exit code
which makes it out is the exit code of the process. Also changing names
to be internally consistent and trying to prune pieces which don't make
so much sense without sbt around. Started on documentation. No review.