Changes to startup.
default. The -toolcp option now works. -D and -J options are passed
as arguments to the runner program as well as to the underlying JVM,
because they may be needed for such things as fsc talking to a server
compiler. Added a new Settings type for -D and -J options. Deleted some
dead settings code.

There are dummy settings for -toolcp and -nobootcp so they can be seen
in the help output, which needs a big cleanup.

No review.