Altered a check for impl classes to look at the...
Altered a check for impl classes to look at the name instead of the
flags, since the flags are wrong. Obviously we should fix the flags or
otherwise better address. Note that the test is well intentioned but
doesn't actually fail with older versions; the crash is not easy to
reproduce even when I simulate the whole repl. Review by dragos.