A small addition to the library to address some...
A small addition to the library to address something bugging me forever.
It's a light interface to system properties. It's not intended to solve
all property issues for all time, only to greatly improve on the overly
ad-hoc ways things are presently done. Feedback welcome. Sorry it's
coming in this late but it arises from writing the tools to fix the bugs
to allow that release to happen. That's nature's circle of bugs. Review
by community.