Various chronic annoyances with the repl addres...
Various chronic annoyances with the repl addressed. Much improved
transcript pasting. Now goes back in time to fix the transcript if it
contains self-referential "res0, res1" etc. so that it works as it
originally did. Shows which commands it is running, and places the
commands with their result in a manner suitable for framing.

Also, a new :paste command which accepts input up to ctrl-D, so you can
enter companions without gyrations, or code from people who write in a
repl unfriendly fashion by putting their curly braces on the next line
(I'm looking at you mark harrah) or you name it, it's not picky.

No review.